+++ src/app/[locale]/viewer/registry/_components/RegistryListView.tsx
@@ -4,7 +4,7 @@ import { useSearchParams } from 'next/navigation';
 import AssetNotFound from 'components/basics/AssetNotFound';
 import { encodeBase64 } from 'lib/util/Base64Util';
 import ListHeader from 'components/basics/ListHeader';
-import { performSearchAasFromAllRepositories } from 'lib/services/repository-access/repositorySearchActions';
+import { performSearchAasFromAllRepositories } from 'lib/services/aas-repository-service/repositorySearchActions';
 import { useTranslations } from 'next-intl';
 import { LocalizedError } from 'lib/util/LocalizedError';
 import { AasListEntry } from 'lib/types/AasListEntry';
