+++ src/lib/hooks/UseAasDataLoader.ts
@@ -1,7 +1,6 @@
 import { useEffect } from 'react';
 import { Reference, Submodel } from 'lib/api/aas/models';
 import { useEnv } from 'app/EnvProvider';
-import { getAasFromRepository } from 'lib/services/search-actions/searchActions';
 import { LocalizedError } from 'lib/util/LocalizedError';
 import { CurrentAasContextType, SubmodelOrIdReference } from 'components/contexts/CurrentAasContext';
 import { SubmodelDescriptor } from 'lib/types/registryServiceTypes';
@@ -10,6 +9,7 @@ import { useAsyncEffect } from 'lib/hooks/UseAsyncEffect';
 import { useAasStore } from 'stores/AasStore';
 import { performFullAasSearch } from 'lib/services/infrastructure-search-service/infrastructureSearchActions';
 import { performSubmodelFullSearch } from 'lib/services/submodel-repository-service/submodelRepositoryActions';
+import { getAasFromRepository } from 'lib/services/aas-repository-service/repositorySearchActions';
 
 /**
  * Hook to load AAS content and its submodels asynchronously.
