+++ src/lib/services/aas-repository-service/repositorySearchActions.ts
@@ -1,14 +1,14 @@
 'use server';
 
 import { AssetAdministrationShell } from 'lib/api/aas/models';
-import { RepoSearchResult, RepositorySearchService } from 'lib/services/repository-access/RepositorySearchService';
+import { RepoSearchResult, RepositorySearchService } from 'lib/services/aas-repository-service/RepositorySearchService';
 import {
     ApiFileDto,
     ApiResponseWrapper,
     wrapErrorCode,
     wrapFile,
 } from 'lib/util/apiResponseWrapper/apiResponseWrapper';
-import { AssetAdministrationShellRepositoryApi, SubmodelRepositoryApi } from 'lib/api/basyx-v3/api';
+import { AssetAdministrationShellRepositoryApi } from 'lib/api/basyx-v3/api';
 import { mnestixFetch } from 'lib/api/infrastructure';
 import { headers } from 'next/headers';
 import { createRequestLogger, logInfo } from 'lib/util/Logger';
@@ -19,7 +19,7 @@ export async function performSearchAasFromAllRepositories(
     const logger = createRequestLogger(await headers());
     logInfo(logger, 'performSearchAasFromAllRepositories', 'Requested AAS', { requestedId: searchInput });
     const searcher = RepositorySearchService.create(logger);
-    return searcher.getFromAllRepos(searchInput);
+    return searcher.searchInAllAasRepositories(searchInput);
 }
 
 export async function getThumbnailFromShell(
@@ -30,29 +30,6 @@ export async function getThumbnailFromShell(
     const searchResponse = await fileSearcher.getThumbnailFromShell(aasId);
     if (!searchResponse.isSuccess) return wrapErrorCode(searchResponse.errorCode, searchResponse.message);
     return wrapFile(searchResponse.result);
-}
-
-export async function getAttachmentFromSubmodelElement(
-    submodelId: string,
-    submodelElementPath: string,
-    baseRepositoryUrl?: string,
-): Promise<ApiResponseWrapper<ApiFileDto>> {
-    const logger = createRequestLogger(await headers());
-    logInfo(logger, getAttachmentFromSubmodelElement.name, 'Requested Attachment', {
-        submodelId: submodelId,
-        submodelElementPath: submodelElementPath,
-    });
-    const searcher = RepositorySearchService.create(logger);
-    if (baseRepositoryUrl) {
-        const fileSearcher = SubmodelRepositoryApi.create(baseRepositoryUrl, mnestixFetch());
-        const searchResponse = await fileSearcher.getAttachmentFromSubmodelElement(submodelId, submodelElementPath);
-        if (!searchResponse.isSuccess) return wrapErrorCode(searchResponse.errorCode, searchResponse.message);
-        return wrapFile(searchResponse.result);
-    }
-
-    const response = await searcher.getFirstAttachmentFromSubmodelElementFromAllRepos(submodelId, submodelElementPath);
-    if (!response.isSuccess) return wrapErrorCode(response.errorCode, response.message);
-    return wrapFile(response.result.searchResult);
 }
 
 export async function downloadAasFromRepo(
@@ -69,4 +46,17 @@ export async function downloadAasFromRepo(
     const response = await fileSearcher.downloadAAS(aasId, submodelIds, includeConceptDescriptions);
     if (!response.isSuccess) return wrapErrorCode(response.errorCode, response.message);
     return response;
+}
+
+export async function getAasFromRepository(
+    aasId: string,
+    repositoryUrl: string,
+): Promise<ApiResponseWrapper<AssetAdministrationShell>> {
+    const logger = createRequestLogger(await headers());
+    logInfo(logger, 'getAasFromRepository', 'Requested AAS/AssetID', {
+        requestedId: aasId,
+        repositoryUrl: repositoryUrl,
+    });
+    const api = AssetAdministrationShellRepositoryApi.create(repositoryUrl, mnestixFetch());
+    return api.getAssetAdministrationShellById(aasId);
 }
